new_bot/requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.0.0

# Telegram bot
python-telegram-bot>=20.7

# Environment variables
python-dotenv>=1.0.0
